requirements.txt
Flask==2.2.2
Flask-Session==0.4.0
pymongo==4.3.3
psycopg2-binary==2.9.5
Flask==2.2.2
Flask-Session==0.4.0
pymongo==4.3.3
psycopg2-binary==2.9.5
